Fixes some test issues with setuptools.
Fixes a tox config issue, where we are unable to import `pofh` in our tests
because the source package is on the path (changedir fixes this).

Makes pytest the default `setup.py test` command, since it can be time consuming
to set up a functioning environment for tox with multiple python versions.

Fixes some failing tests in the test_sms_dispatcher test module.